+++ packages/ui/src/solidity/ERC20Controls.svelte
@@ -202,6 +202,11 @@
 
 <AccessControlSection bind:access={opts.access} required={requireAccessControl} />
 
-<UpgradeabilitySection bind:upgradeable={opts.upgradeable} />
+<UpgradeabilitySection
+  bind:upgradeable={opts.upgradeable}
+  namespaceRequired={opts.upgradeable !== false && opts.crossChainBridging === 'custom'}
+  bind:namespacePrefix={opts.namespacePrefix}
+  {errors}
+/>
 
 <InfoSection bind:info={opts.info} />